fixed a 64bit issue
2009-05-02  Jeffrey Stedfast  <fejj@novell.com>

	* gmime/gmime-stream-buffer.c (stream_seek_cache_read): Fixed a
	64bit cleanlines issue.

+++ gmime/gmime-stream-buffer.c
@@ -489,7 +489,7 @@ static gint64
 stream_seek_cache_read (GMimeStream *stream, gint64 offset, GMimeSeekWhence whence)
 {
 	GMimeStreamBuffer *buffer = (GMimeStreamBuffer *) stream;
-	size_t len, total = 0;
+	gint64 len, total = 0;
 	gint64 pos, real;
 	ssize_t nread;
